[CALCITE-3550] Make SqlTypeAssignmentRules conversion mapping pluggable
* Add an utility class named SqlTypeMappingRules that defines some utilities
to build type mappings;
* Add an interface named SqlTypeMappingRule for all the type mapping
rules;
* Add SqlTypeCoercionRule that defines the type coercion rules, add doc
about how to customize new rules;
* Rename SqlTypeAssignmentRules to SqlTypeAssignmentRule;
* SqlTypeAssignmentRule now only defines rules of type assignment;
* Add a new method SqlValidator.setSqlTypeCoercionRules to set up
custom type coercion rules.
