+++ core/src/main/java/org/apache/calcite/sql/validate/SqlValidatorImpl.java
@@ -84,6 +84,7 @@ import org.apache.calcite.sql.parser.SqlParserPos;
 import org.apache.calcite.sql.type.AssignableOperandTypeChecker;
 import org.apache.calcite.sql.type.ReturnTypes;
 import org.apache.calcite.sql.type.SqlOperandTypeInference;
+import org.apache.calcite.sql.type.SqlTypeCoercionRule;
 import org.apache.calcite.sql.type.SqlTypeName;
 import org.apache.calcite.sql.type.SqlTypeUtil;
 import org.apache.calcite.sql.util.SqlBasicVisitor;
@@ -3886,6 +3887,10 @@ public class SqlValidatorImpl implements SqlValidatorWithHints {
   public TypeCoercion getTypeCoercion() {
     assert isTypeCoercionEnabled();
     return this.typeCoercion;
+  }
+
+  public void setSqlTypeCoercionRules(SqlTypeCoercionRule typeCoercionRules) {
+    SqlTypeCoercionRule.THREAD_PROVIDERS.set(typeCoercionRules);
   }
 
   /**
